### 2022-08-20 #4688 @LTLA
Thanks @arfon. If it helps to make a decision, this work also includes several utility packages that I didn't mention explicitly in the manuscript:

- https://github.com/LTLA/basilisk.utils
- https://github.com/LTLA/dir.expiry

For various reasons, they needed to be split into separate packages rather than remaining part of **basilisk** itself. I'd guess that **basilisk.utils** would probably add a few hundred lines of extra code, if it were included.
