### 2022-08-25 9cac618
Switched dummy license to GPL-3 to keep Legal happy.
