+++ inst/example/inst/test_dummy/setup.py
@@ -6,7 +6,7 @@ setup(name='test_dummy',
       url='http://github.com/LTLA/basilisk',
       author='Aaron Lun',
       author_email='infinite.monkeys.with.keyboards@gmail.com',
-      license='MIT',
+      license='GPL-3',
       packages=['test_dummy'],
       zip_safe=False)
 
