+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: basilisk
-Version: 1.8.0
-Date: 2021-04-03
+Version: 1.8.1
+Date: 2022-08-24
 Title: Freezing Python Dependencies Inside Bioconductor Packages
 Authors@R: c(person("Aaron", "Lun", role=c("aut", "cre", "cph"),
         email="infinite.monkeys.with.keyboards@gmail.com"),
