### 2022-08-25 7b2c3c7
Switched dummy license to GPL-3 to keep Legal happy.
