+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: basilisk
-Version: 1.9.2
-Date: 2022-06-23
+Version: 1.9.3
+Date: 2022-08-24
 Title: Freezing Python Dependencies Inside Bioconductor Packages
 Authors@R: c(person("Aaron", "Lun", role=c("aut", "cre", "cph"),
         email="infinite.monkeys.with.keyboards@gmail.com"),
