### 2022-08-27 #4688 @arfon
:wave: @Nikoleta-v3 – would you be willing to edit this submission for JOSS? I realize that bioinformatics isn't (even close to) your field, but this submission looks to be more about how to leverage Python tooling from an R interface which does seem like something you might be able to edit?
