+++ ref.bib
@@ -32,6 +32,7 @@
   title = {A comprehensive {B}ioconductor ecosystem for the design of CRISPR guide RNAs across nucleases and technologies},
   year = {2022},
   URL = {https://www.biorxiv.org/content/early/2022/04/28/2022.04.21.488824},
+  doi = {10.1101/2022.04.21.488824},
   journal = {bioRxiv}
 }
 
