### 2022-08-30 #4688 @LTLA
From the list, I'd say that **rmflight**, **nhejazi**, **mschubert** and **cbrueffer** could be appropriate reviewers. (Not excluding anyone here, I just searched for "R", "Python" and "bioinformatics" and picked a few names randomly.)

I see **alanocallaghan** on the list - he is already the maintainer of several **basilisk** clients. This may or may not be a desirable characteristic of a reviewer.
