### 2022-08-31 #4688 @Nikoleta-v3
👋🏻  @rmflight @mschubert @cbrueffer would any of you be willing to review this submission for JOSS?

The submission I'd like you to review is titled: "basilisk: a Bioconductor package for managing Python environments". You can find more information at the top of this Github issue 🆙
