### 2022-08-31 8e36de8
Update examples and tests to work with a newer Python.
