+++ R/useBasiliskEnv.R
@@ -21,10 +21,9 @@
 #' @examples
 #' \dontshow{basilisk.utils::installConda()}
 #'
-#' tmploc <- file.path(tempdir(), "my_package_B")
+#' tmploc <- file.path(tempdir(), "my_package_A")
 #' if (!file.exists(tmploc)) {
-#'     setupBasiliskEnv(tmploc, c('pandas==0.25.1',
-#'         "python-dateutil=2.8.0", "pytz=2019.3"))
+#'     setupBasiliskEnv(tmploc, c('pandas==1.4.3'))
 #' }
 #'
 #' # This may or may not work, depending on whether a Python instance
