+++ tests/testthat/setup.R
@@ -1,10 +1,10 @@
 # Setting up common variables.
 
-test.pandas <- "pandas==0.25.1"
-test.pandas.deps <- c("python-dateutil==2.8.0", "pytz==2019.3")
+test.pandas <- "pandas==1.4.3"
+test.pandas.deps <- c("python-dateutil==2.8.2", "pytz==2022.2.1")
 
-old.pandas <- "pandas==0.24.1"
-old.pandas.deps <- c("python-dateutil==2.7.1", "pytz==2018.7")
+old.pandas <- "pandas==1.4.2"
+old.pandas.deps <- c("python-dateutil==2.8.1", "pytz==2022.1")
 
 client.dir <- "install-test-client"
 unlink(client.dir, recursive=TRUE)
