+++ man/setupBasiliskEnv.Rd
@@ -76,8 +76,7 @@ If no Python version is listed, the version in the base conda installation is us
 
 tmploc <- file.path(tempdir(), "my_package_A")
 if (!file.exists(tmploc)) {
-    setupBasiliskEnv(tmploc, c('pandas=0.25.3',
-        "python-dateutil=2.8.1", "pytz=2019.3"))
+    setupBasiliskEnv(tmploc, c('pandas=1.4.3'))
 }
 
 }
