### 2022-09-03 50b3315
Use the internal R for fallback execution on GLIBCXX error.
Added testload= to check whether the fallback is necessary.
