+++ man/basiliskStart.Rd
@@ -6,7 +6,12 @@
 \alias{basiliskRun}
 \title{Start and stop \pkg{basilisk}-related processes}
 \usage{
-basiliskStart(env, fork = getBasiliskFork(), shared = getBasiliskShared())
+basiliskStart(
+  env,
+  fork = getBasiliskFork(),
+  shared = getBasiliskShared(),
+  testload = NULL
+)
 
 basiliskStop(proc)
 
@@ -16,7 +21,8 @@ basiliskRun(
   ...,
   env,
   fork = getBasiliskFork(),
-  shared = getBasiliskShared()
+  shared = getBasiliskShared(),
+  testload = NULL
 )
 }
 \arguments{
@@ -32,6 +38,9 @@ If \code{FALSE}, a new worker process is created using communication over socket
 
 \item{shared}{Logical scalar indicating whether \code{basiliskStart} is allowed 
 to load a shared Python instance into the current R process, see \code{\link{getBasiliskShared}}.}
+
+\item{testload}{Character vector specifying the Python packages to load into the process during set-up.
+This is used to check that packages can be correctly loaded, switching to a fallback on \code{GLIBCXX} dynamic linking failures.}
 
 \item{proc}{A process object generated by \code{basiliskStart}.}
 
@@ -71,7 +80,8 @@ If \code{proc=NULL} in \code{basiliskRun}, a process will be created and closed 
 This may be convenient in functions where persistence is not required.
 Note that doing so requires specification of \code{pkgname} and \code{envname}.
 
-If the base Conda installation provided with \pkg{basilisk} satisfies the requirements of the client package, developers can set \code{env=NULL} in this function to use that base installation rather than constructing a separate environment.
+If the base Conda installation provided with \pkg{basilisk} satisfies the requirements of the client package, 
+developers can set \code{env=NULL} in this function to use that base installation rather than constructing a separate environment.
 }
 \section{Choice of process type}{
 
@@ -95,19 +105,33 @@ Developers can control these choices directly by explicitly specifying \code{sha
 while users can control them indirectly with \code{\link{setBasiliskFork}} and related functions.
 }
 
+\section{Testing package loads}{
+
+If \code{testload} is provided, \code{basiliskStart} will attempt to load those Python packages into the process.
+This is used to detect loading failures due to differences in the versions of the shared libraries.
+Most typically, a conda-supplied Python package (often \pkg{scipy} submodules) will have been compiled against a certain version of \code{libstdc++},
+but R is compiled against an older version, and R's version takes precedence when the package is loaded via \pkg{reticulate}.
+
+If loading failures are encountered, \code{basiliskStart} will fall back to a separate socket process running a Conda-supplied R installation.
+The hope is that, if both Python and R are sourced from conda, they will be using the same version of \code{libstdc++}. 
+When used, this fallback overrides any choice of process type  from \code{fork} and \code{shared}.
+}
+
 \section{Constraints on user-defined functions}{
 
 In \code{basiliskRun}, there is no guarantee that \code{fun} has access to \code{basiliskRun}'s calling environment.
 This has a number of consequences for the type of code that can be written inside \code{fun}:
 \itemize{
-\item Functions or variables from non-base R packages used inside \code{fun} should be prefixed with the package namespace, 
-or the package itself should be reloaded inside \code{fun}.
 \item Any other variables used inside \code{fun} should be explicitly passed as an argument.
 Developers should not rely on closures to capture variables in the calling environment of \code{basiliskRun}.
 \item Relevant global variables from the calling environment should be explicitly reset inside \code{fun}.
 \item Developers should \emph{not} attempt to pass complex objects to memory in or out of \code{fun}.
[…]
