+++ NAMESPACE
@@ -19,6 +19,7 @@ export(useBasiliskEnv)
 exportClasses(BasiliskEnvironment)
 import(basilisk.utils)
 import(methods)
+importFrom(basilisk.utils,activateEnvironment)
 importFrom(basilisk.utils,getSystemDir)
 importFrom(dir.expiry,lockDirectory)
 importFrom(dir.expiry,touchDirectory)
