### 2022-09-04 59f34fb
Documented the limitations of the internal R fallback.
