### 2022-09-04 25f29b5
Added option to force the use of the fallback for easier testing.
