+++ NAMESPACE
@@ -7,11 +7,13 @@ export(basiliskStart)
 export(basiliskStop)
 export(configureBasiliskEnv)
 export(findPersistentEnv)
+export(getBasiliskForceFallback)
 export(getBasiliskFork)
 export(getBasiliskShared)
 export(listCorePackages)
 export(listPackages)
 export(obtainEnvironmentPath)
+export(setBasiliskForceFallback)
 export(setBasiliskFork)
 export(setBasiliskShared)
 export(setupBasiliskEnv)

[…]
