+++ vignettes/motivation.Rmd
@@ -188,6 +188,10 @@ However, it is not desirable if there are multiple R instances running different
 (Multiple R instances running different Bioconductor releases are not affected.)
 This option has no effect if `BASILISK_USE_SYSTEM_DIR` is set.
 
+Setting `BASILISK_NO_CONDA_R` to `1` will instruct `r Biocpkg("basilisk")` to _not_ create an internal conda-based R installation for the last-resort fallback (see comments in `?basiliskStart`).
+This avoids the overhead of an internal R installation if it is known that shared library version conflicts will not occur.
+It is most useful for streamlining applications where developers can (i) test that no conflicts occur in the subset of `r Biocpkg("basilisk")` clients and/or (ii) control the versions of system libraries used by R. 
+
 ## Custom environment creation
 
 Given a path, `basiliskStart()` can operate on any conda or virtual environment, not just those constructed by `r Biocpkg("basilisk")`.
