### 2022-09-04 915c34c
Only create the fallback environment on demand.
