+++ NAMESPACE
@@ -22,6 +22,7 @@ exportClasses(BasiliskEnvironment)
 import(basilisk.utils)
 import(methods)
 importFrom(basilisk.utils,activateEnvironment)
+importFrom(basilisk.utils,getFallbackREnv)
 importFrom(basilisk.utils,getSystemDir)
 importFrom(dir.expiry,lockDirectory)
 importFrom(dir.expiry,touchDirectory)
