### 2022-09-04 ef04473
Fixed example to ensure conda installation.
