+++ R/obtainEnvironmentPath.R
@@ -13,6 +13,8 @@
 #' @author Aaron Lun
 #'
 #' @examples
+#' \dontshow{basilisk.utils::installConda()}
+#'
 #' tmploc <- file.path(tempdir(), "my_package_A")
 #' if (!file.exists(tmploc)) {
 #'     setupBasiliskEnv(tmploc, c('pandas=1.4.3'))
