+++ R/basiliskStart.R
@@ -16,6 +16,8 @@
 #' to load a shared Python instance into the current R process, see \code{\link{getBasiliskShared}}.
 #' @param fun A function to be executed in the \pkg{basilisk} process.
 #' This should return a \dQuote{pure R} object, see details.
+#' @param testload Character vector specifying the Python packages to load into the process during set-up.
+#' This is used to check that packages can be correctly loaded, switching to a fallback on \code{GLIBCXX} dynamic linking failures. 
 #' @param ... Further arguments to be passed to \code{fun}.
 #'
 #' @return 
@@ -45,7 +47,8 @@
 #' This may be convenient in functions where persistence is not required.
 #' Note that doing so requires specification of \code{pkgname} and \code{envname}.
 #' 
-#' If the base Conda installation provided with \pkg{basilisk} satisfies the requirements of the client package, developers can set \code{env=NULL} in this function to use that base installation rather than constructing a separate environment.
+#' If the base Conda installation provided with \pkg{basilisk} satisfies the requirements of the client package, 
+#' developers can set \code{env=NULL} in this function to use that base installation rather than constructing a separate environment.
 #'
 #' @section Choice of process type:
 #' \itemize{
@@ -67,19 +70,40 @@
 #' Developers can control these choices directly by explicitly specifying \code{shared} and \code{fork},
 #' while users can control them indirectly with \code{\link{setBasiliskFork}} and related functions.
 #'
+#' @section Testing package loads:
+#' If \code{testload} is provided, \code{basiliskStart} will attempt to load those Python packages into the newly created process.
+#' This is used to detect loading failures due to differences in the versions of the shared libraries.
+#' Most typically, a conda-supplied Python package (often \pkg{scipy} submodules) will have been compiled against a certain version of \code{libstdc++} but R is compiled against an older version.
+#' R's version takes precedence when \pkg{reticulate} attempts to load the Python package, causing cryptic \dQuote{GLIBCXX version not found} errors.
+#' 
+#' By checking the specified \code{testload}, \code{basiliskStart} can check for loading failures in potentially problematic packages.
+#' Upon any failure, \code{basiliskStart} will fall back to a separate socket process running a conda-supplied R installation.
+#' The idea is that, if both Python and R are sourced from conda, they will be using the same version of \code{libstdc++} and other libraries.
+#' This avoids loading errors and/or segmentation faults due to version mismatches.
+#'
+#' Use of this "last resort fallback" overrides any choice of process type from \code{fork} and \code{shared}.
+#' If no failures are encountered, a process will be created using the current R installation.
+#'
+#' Note that the fallback R installation is very minimalistic; only \pkg{reticulate} is guaranteed to be available.
+#' This places some limitations on the code that can be executed inside \code{fun} for \pkg{basilisk} environments that might trigger use of the fallback.
+#'
 #' @section Constraints on user-defined functions:
 #' In \code{basiliskRun}, there is no guarantee that \code{fun} has access to \code{basiliskRun}'s calling environment.
-#' This has a number of consequences for the type of code that can be written inside \code{fun}:
+#' This has several consequences for code in the body of \code{fun}:
 #' \itemize{
-#' \item Functions or variables from non-base R packages used inside \code{fun} should be prefixed with the package namespace, 
-#' or the package itself should be reloaded inside \code{fun}.
[…]
