+++ NAMESPACE
@@ -7,11 +7,13 @@ export(basiliskStart)
 export(basiliskStop)
 export(configureBasiliskEnv)
 export(findPersistentEnv)
+export(getBasiliskForceFallback)
 export(getBasiliskFork)
 export(getBasiliskShared)
 export(listCorePackages)
 export(listPackages)
 export(obtainEnvironmentPath)
+export(setBasiliskForceFallback)
 export(setBasiliskFork)
 export(setBasiliskShared)
 export(setupBasiliskEnv)
@@ -19,6 +21,8 @@ export(useBasiliskEnv)
 exportClasses(BasiliskEnvironment)
 import(basilisk.utils)
 import(methods)
+importFrom(basilisk.utils,activateEnvironment)
+importFrom(basilisk.utils,getFallbackREnv)
 importFrom(basilisk.utils,getSystemDir)
 importFrom(dir.expiry,lockDirectory)
 importFrom(dir.expiry,touchDirectory)

[…]
