+++ R/basiliskStart.R
@@ -16,6 +16,8 @@
 #' to load a shared Python instance into the current R process, see \code{\link{getBasiliskShared}}.
 #' @param fun A function to be executed in the \pkg{basilisk} process.
 #' This should return a \dQuote{pure R} object, see details.
+#' @param persist Logical scalar indicating whether to pass a persistent store to \code{fun}.
+#' If \code{TRUE}, \code{fun} should accept a \code{store} argument.
 #' @param testload Character vector specifying the Python packages to load into the process during set-up.
 #' This is used to check that packages can be correctly loaded, switching to a fallback on \code{GLIBCXX} dynamic linking failures. 
 #' @param ... Further arguments to be passed to \code{fun}.
@@ -33,11 +35,6 @@
 #' This avoids version conflicts in the presence of other Python instances or environments loaded by other packages or by the user.
 #' Thus, \pkg{basilisk} clients are not affected by (and if \code{shared=FALSE}, do not affect) the activity of other R packages.
 #' 
-#' If necessary, objects created in \code{fun} can persist across calls to \code{basiliskRun}, e.g., for file handles.
-#' This requires the use of \code{\link{assign}} with \code{envir} set to \code{\link{findPersistentEnv}} to persist a variable,
-#' and a corresponding \code{\link{get}} to retrieve that object in later calls. 
-#' See Examples for more details.
-#'
 #' It is good practice to call \code{basiliskStop} once computation is finished to terminate the process.
 #' Any Python-related operations between \code{basiliskStart} and \code{basiliskStop} should only occur via \code{basiliskRun}.
 #' Calling \pkg{reticulate} functions directly will have unpredictable consequences,
@@ -105,6 +102,17 @@
 #' If \code{fun} involves fallback-inducing Python packages, developers can further set \code{\link{setBasiliskForceFallback}} before running \code{basiliskRun}.
 #' This tests that \code{fun} works with the minimal conda-supplied R installation.
 #'
+#' @section Persisting objects across calls:
+#' Objects created inside \code{fun} can be persisted across calls to \code{basiliskRun} by setting \code{persist=TRUE}.
+#' This will instruct \code{basiliskRun} to pass a \code{store} argument to \code{fun} that can be used to store arbitrary objects.
+#' Those same objects can be retrieved from \code{store} in later calls to \code{basiliskRun} using the same \code{proc}.
+#' Any object can be stored in \code{.basilisk.store} but will remain strictly internal to \code{proc}.
+#'
+#' This capability is primarily useful when a Python workflow is split across multiple \code{basiliskRun} calls.
+#' Each subsequent call can pick up from temporary intermediate objects generated by the previous call.
+#' In this manner, \pkg{basilisk} enables modular function design where developers can easily mix and match different \code{basiliskRun} invocations.
+#' See Examples for a working demonstration.
+#'
 #' @section Use of lazy installation:
 #' If the specified \pkg{basilisk} environment is not present and \code{env} is a \linkS4class{BasiliskEnvironment} object, the environment will be created upon first use of \code{basiliskStart}.
 #' If the base Conda installation is not present, it will also be installed upon first use of \code{basiliskStart}.
@@ -168,12 +176,13 @@
 #'
[…]
