+++ man/basiliskStart.Rd
@@ -20,6 +20,7 @@ basiliskRun(
   fun,
   ...,
   env,
+  persist = FALSE,
   fork = getBasiliskFork(),
   shared = getBasiliskShared(),
   testload = NULL
@@ -48,6 +49,9 @@ This is used to check that packages can be correctly loaded, switching to a fall
 This should return a \dQuote{pure R} object, see details.}
 
 \item{...}{Further arguments to be passed to \code{fun}.}
+
+\item{persist}{Logical scalar indicating whether to pass a persistent store to \code{fun}.
+If \code{TRUE}, \code{fun} should accept a \code{store} argument.}
 }
 \value{
 \code{basiliskStart} returns a process object, the exact nature of which depends on \code{fork} and \code{shared}.
@@ -65,11 +69,6 @@ can be safely performed with the correct versions of Python packages.
 These functions ensure that any Python operations in \code{fun} will use the environment specified by \code{envname}.
 This avoids version conflicts in the presence of other Python instances or environments loaded by other packages or by the user.
 Thus, \pkg{basilisk} clients are not affected by (and if \code{shared=FALSE}, do not affect) the activity of other R packages.
-
-If necessary, objects created in \code{fun} can persist across calls to \code{basiliskRun}, e.g., for file handles.
-This requires the use of \code{\link{assign}} with \code{envir} set to \code{\link{findPersistentEnv}} to persist a variable,
-and a corresponding \code{\link{get}} to retrieve that object in later calls. 
-See Examples for more details.
 
 It is good practice to call \code{basiliskStop} once computation is finished to terminate the process.
 Any Python-related operations between \code{basiliskStart} and \code{basiliskStop} should only occur via \code{basiliskRun}.
@@ -144,6 +143,19 @@ If \code{fun} involves fallback-inducing Python packages, developers can further
 This tests that \code{fun} works with the minimal conda-supplied R installation.
 }
 
+\section{Persisting objects across calls}{
+
+Objects created inside \code{fun} can be persisted across calls to \code{basiliskRun} by setting \code{persist=TRUE}.
+This will instruct \code{basiliskRun} to pass a \code{store} argument to \code{fun} that can be used to store arbitrary objects.
+Those same objects can be retrieved from \code{store} in later calls to \code{basiliskRun} using the same \code{proc}.
+Any object can be stored in \code{.basilisk.store} but will remain strictly internal to \code{proc}.
+
+This capability is primarily useful when a Python workflow is split across multiple \code{basiliskRun} calls.
+Each subsequent call can pick up from temporary intermediate objects generated by the previous call.
+In this manner, \pkg{basilisk} enables modular function design where developers can easily mix and match different \code{basiliskRun} invocations.
+See Examples for a working demonstration.
+}
+
 \section{Use of lazy installation}{
 
 If the specified \pkg{basilisk} environment is not present and \code{env} is a \linkS4class{BasiliskEnvironment} object, the environment will be created upon first use of \code{basiliskStart}.
@@ -204,12 +216,13 @@ basiliskStop(cl2)
 
 # Persistence of variables is possible within a Start/Stop pair.
 cl <- basiliskStart(tmploc)
-basiliskRun(proc=cl, function() {
-    assign(x="snake.in.my.shoes", 1, envir=basilisk::findPersistentEnv())
-})
-basiliskRun(proc=cl, function() {
-    get("snake.in.my.shoes", envir=basilisk::findPersistentEnv())
-})
+basiliskRun(proc=cl, function(store) {
+    store$snake.in.my.shoes <- 1
+    invisible(NULL)
+}, persist=TRUE)
+basiliskRun(proc=cl, function(store) {
+    return(store$snake.in.my.shoes)
+}, persist=TRUE)
 basiliskStop(cl)
 
 }

[…]
