### 2022-09-05 7a4e501
Fixes to keep check happy, added tests for the fallback.
