### 2022-09-06 #4688 @Nikoleta-v3
Thank you for letting me know about this potential competing interest @gtonkinhill 🙏🏻  if you feel that you can objectively evaluate the paper, then I see no reason why you shouldn't be a reviewer 😄
