### 2022-09-07 0cb335c
Fixed the testload for fallback in the example package.
