+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: basilisk
-Version: 1.9.5
-Date: 2022-09-02
+Version: 1.9.6
+Date: 2022-09-06
 Title: Freezing Python Dependencies Inside Bioconductor Packages
 Authors@R: c(person("Aaron", "Lun", role=c("aut", "cre", "cph"),
         email="infinite.monkeys.with.keyboards@gmail.com"),
