### 2022-09-07 #4742 @editorialbot
[REVIEW]: basilisk: a Bioconductor package for managing Python environments
**Submitting author:** @LTLA (<a href="http://orcid.org/0000-0002-3564-4813">Aaron Lun</a>)
**Repository:** https://github.com/LTLA/basilisk
**Branch with paper.md** (empty if default branch): submission
**Version:** 1.9.11
