### 2022-09-08 #4742 @jsun
Hi, @LTLA, I finished my review. The manuscript and vignette are well-written, following them I can easily try the package. I installed the package and ran some examples and I experienced no issues when using this package. Although this package is intended for use by developers, general users such as data analysts may also be interested. This is because many biological analyses should be performed alternately in R and Python; of course, sometimes require multiple versions of Python. From the viewpoint of a non-developer user, I would suggest that the author add some easy-to-understand sample codes to the vignette. Here are my suggestions.

(1) Although the author mentioned the following items

- ... will lazily install conda and the required environments if they are not already present. ... (vignette 3.4)
- ... It is possible to direct basilisk to use an existing Miniconda or Anaconda instance ... (vignette 4.2)

I suggest the author merged them into one section. For example, add an Installation section to mention the following items.

- basilisk can use existing conda or can install conda automatically
- some caution for environment variables (e.g., BASILISK_USE_SYSTEM_DIR) that should be known/set before using basilisk main functions.

(2) Most codes in the vignette are examples. Is it possible to add some executable codes into the vignette? For example,

- add codes in the manuscript to the vignette. This is because I found that if I prepare `input_matrix` with random numbers, I can execute the code written in the manuscript.
- add an example that just imports sklearn.datasets, loads the iris dataset, and returns it to the R variable.

(3) For the license issue or the personal preference, users may use other Python virtual environments such as pyenv instead of `conda`. The vignette 4.3 mentioned that basilisk can support other virtual environments. Are pyenv, venv, etc, included? Is it possible to add an example of how to use this function?
