### 2022-09-11 8949a8a
Added a section describing how end-users can use basilisk.
