+++ vignettes/motivation.Rmd
@@ -3,7 +3,7 @@ title: Freezing Python versions inside Bioconductor packages
 author: 
 - name: Aaron Lun
   email: infinite.monkeys.with.keyboards@gmail.com
-date: "Revised: January 25, 2019"
+date: "Revised: September 10, 2022"
 output:
   BiocStyle::html_document
 package: basilisk
@@ -59,12 +59,12 @@ These objects define the Python environments to be constructed by `r Biocpkg("ba
 ```{r}
 my_env <- BasiliskEnvironment(envname="my_env_name",
     pkgname="ClientPackage",
-    packages=c("pandas==0.25.1", "sklearn==0.22.1")
+    packages=c("pandas==1.4.3", "scikit-learn==1.1.1")
 )
 
 second_env <- BasiliskEnvironment(envname="second_env_name",
     pkgname="ClientPackage",
-    packages=c("scipy=1.4.1", "numpy==1.17") 
+    packages=c("scipy=1.9.1", "numpy==1.22.1") 
 )
 ```
 
@@ -210,7 +210,7 @@ Similarly, all (correctly `configure`d) client packages will install their envir
 This is very useful for enterprise-level deployments as the conda instances and environments are (i) not duplicated in each user's home directory, and (ii) always available to any user with access to the R installation. 
 However, it requires installation from source and thus is not set by default.
 
-It is possible to direct `r Biocpkg("basilisk")` to use an existing Miniconda or Anaconda instance by setting the `BASILISK_EXTERNAL_CONDA` environment variable to an absolute path to the installation directory.
+It is possible to direct `r Biocpkg("basilisk")` to use an existing Miniconda or Anaconda instance by setting the `BASILISK_EXTERNAL_CONDA` environment variable to an absolute path to the conda installation directory.
 This may be desirable to avoid redundant copies of the same installation.
 However, in such cases, it is the **user's responsibility** to administer their conda instance to ensure that the correct versions of all packages are available for `r Biocpkg("basilisk")`'s clients.
 
@@ -225,12 +225,61 @@ This avoids the overhead of an internal R installation if it is known that share
 It is most useful for streamlining applications where developers can (i) test that no conflicts occur in the subset of `r Biocpkg("basilisk")` clients and/or (ii) control the versions of system libraries used by R.
 This option only has an effect if `BASILISK_USE_SYSTEM_DIR` is set, as otherwise the fallback R installation is only created upon encountering version conflicts.
 
-## Custom environment creation
+# Non-package use
 
-Given a path, `basiliskStart()` can operate on any conda or virtual environment, not just those constructed by `r Biocpkg("basilisk")`.
-It is possible to manually set up the desired environment with, e.g., Python 2.7, or to manually modify an existing environment to include custom Python packages not in PyPi or Conda.
-This allows users to continue to take advantage of `r Biocpkg("basilisk")`'s graceful handling of Python collisions in more complex applications. 
-However, it can be quite difficult to guarantee portability so one should try to avoid doing this.
+While `r Biocpkg("basilisk")` is primarily intended for package developers, end-users can also take advantage of its graceful handling of multiple Python environments in complex workflows.
+For example, we can easily instantiate a conda environment in our working directory with `setupBasiliskEnv()`:
+
+```{r}
+# Environment location should be an absolute path.
+tmp <- file.path(getwd(), "basilisk-vignette-test") 
+
+# Avoid reinstantiating the environment if it already exists, for faster
+# re-execution of a workflow.
+if (!file.exists(tmp)) {
+    setupBasiliskEnv(tmp, packages=c("scikit-learn=1.1.1", "numpy=1.22.1"))
[…]
