+++ vignettes/motivation.Rmd
@@ -19,7 +19,7 @@ library(basilisk)
 library(BiocStyle)
 ```
 
-# Why?
+# Overview 
 
 Packages like `r CRANpkg("reticulate")` facilitate the use of Python modules in our R-based data analyses, allowing us to leverage Python's strengths in fields such as machine learning and image analysis. 
 However, it is notoriously difficult to ensure that a consistent version of Python is available with a consistently versioned set of modules, especially when the system installation of Python is used.
@@ -27,14 +27,12 @@ As a result, we cannot easily guarantee that some Python code executed via `r CR
 It is also possible that two R packages depend on incompatible versions of Python modules, such that it is impossible to use both packages at the same time.
 These versioning issues represent a major obstacle to reliable execution of Python code across a variety of systems via R/Bioconductor packages.
 
-# What?
-
-`r Biocpkg("basilisk")` uses the [conda installers](https://repo.anaconda.com/) to provision a Python instance that is fully managed by the Bioconductor installation machinery.
+`r Biocpkg("basilisk")` uses [conda](https://repo.anaconda.com/) to provision a Python instance that is fully managed by the Bioconductor installation machinery.
 This provides developers of downstream Bioconductor packages with more control over their Python environment, most typically by the creation of package-specific `conda` environments containing all of their required Python packages.
 Additionally, `r Biocpkg("basilisk")` provides utilities to manage different Python environments within a single R session, enabling multiple Bioconductor packages to use incompatible versions of Python packages in the course of a single analysis.
 These features enable reproducible analysis, simplify debugging of code and improve interoperability between compliant packages.
 
-# How?
+# For package developers
 
 ## Overview
 
@@ -173,17 +171,25 @@ my_example_function <- function() {
 ```
 
 This capability allows developers to modularize complex Python workflows by splitting up steps across multiple calls to `basiliskRun()`. 
-However, it is probably unwise to re-use `proc` across user-visible functions, i.e., the end-user should never have an opportunity to interact with `proc`.
+However, it is probably unwise to re-use `proc` across user-visible functions, i.e., the end user should never have an opportunity to interact with `proc`.
 
-# Additional notes
+# For end users
 
-## Important restrictions  
+## Troubleshooting known issues
+
+In most cases, end users should not have to read this document.
+Properly configured `r Biocpkg("basilisk")` clients should handle all aspects of Python environment creation and loading without requiring user intervention.
+That said, some system configurations are less cooperative than others:
+this section contains a list of known issues and possible fixes.
 
 The Miniconda/Anaconda installers do not work if the installation directory contains spaces.
-If this is the case, consider setting the `BASILISK_EXTERNAL_DIR` environment variable (see below).
+If this is the case for the default installation directory (check `basilisk.utils::getExternalDir()`), 
+consider setting the [`BASILISK_EXTERNAL_DIR`](#fine-tuning-basilisks-behavior) environment variable. 
 
-Windows has a limit of 260 characters on its file paths;
-this can be exceeded due to deeply nested directories for some packages, causing the installation to be incomplete or fail outright.
+Windows has a limit of 260 characters for its file paths.
+This is occasionally exceeded due to deeply nested directories for some packages, causing the conda installation to be incomplete or fail outright.
+Sometimes this can be circumvented by setting `BASILISK_EXTERNAL_DIR` to a shorter path and hoping that enough space is left for the environments.
+
 Builds for 32-bit Windows are not supported due to a lack of demand relative to the difficulty of setting it up.
 
[…]
