### 2022-09-11 1dc4b34
Ensure that the conda environment is actually installed for manual set-up.
