+++ vignettes/motivation.Rmd
@@ -208,9 +208,8 @@ Administrators of an R installation can modify the behavior of `r Biocpkg("basil
 All environment variables described here must be set at both installation time and run time to have any effect.
 If any value is changed, it is generally safest to reinstall `r Biocpkg("basilisk")` and all of its clients.
 
-Setting the `BASILISK_EXTERNAL_DIR` environment variable will change where the conda instance and environments are placed by `basiliskStart()` during lazy installation.
-This is usually unnecessary unless the default path generated by `r CRANpkg("rappdirs")` contains spaces
-or the combination of the default location and conda's directory structure exceeds the file path length limit on Windows.
+Setting the `BASILISK_EXTERNAL_DIR` environment variable will change where the conda instance and environments are placed by [`basiliskStart()`](#basics) during lazy installation.
+This is usually unnecessary unless the default path contains spaces or the combination of the default location and conda's directory structure exceeds the file path length limit on Windows.
 
 Setting `BASILISK_USE_SYSTEM_DIR` to `1` will instruct `r Biocpkg("basilisk")` to install the conda instance in the R system directory during R package installation.
 Similarly, all (correctly `configure`d) client packages will install their environments in the corresponding system directory when they themselves are being installed.
@@ -244,6 +243,7 @@ tmp <- file.path(getwd(), "basilisk-vignette-test")
 # Avoid reinstantiating the environment if it already exists, for faster
 # re-execution of a workflow.
 if (!file.exists(tmp)) {
+    basilisk.utils::installConda() # just in case it hasn't been run before.
     setupBasiliskEnv(tmp, packages=c("scikit-learn=1.1.1", "numpy=1.22.1"))
 }
 ```
@@ -252,7 +252,7 @@ We can then supply this environment's path to `basiliskRun()` to execute Python-
 To demonstrate, we'll apply `r PyPiLink("scikit-learn")`'s truncated PCA on a random matrix.
 Note that the restrictions mentioned [above](#function-constraints) for `fun` are still applicable here.
 
-```{r, error=FALSE}
+```{r, error=FALSE, message=FALSE}
 x <- matrix(rnorm(1000), ncol=10)
 basiliskRun(env=tmp, fun=function(mat) {
     module <- reticulate::import("sklearn.decomposition")
@@ -266,7 +266,7 @@ basiliskRun(env=tmp, fun=function(mat) {
 Of course, in this case, it is the user's responsibility to ensure that the environment is correctly provisioned.
 Some caution is also required as `r Biocpkg("basilisk")` is not guaranteed to work with environments generated by a different version of conda, though problems seem to be rare in practice.
 
-```{r, error=FALSE}
+```{r, error=FALSE, message=FALSE}
 library(reticulate)
 
 # Using reticulate directly to construct our conda environment; though we'll
@@ -274,7 +274,10 @@ library(reticulate)
 # will try to install its own miniconda.
 tmp2 <- file.path(getwd(), "basilisk-vignette-test2")
 if (!file.exists(tmp2)) {
-    conda.bin <- file.path(basilisk.utils::getCondaDir(), basilisk.utils::getCondaBinary())
+    conda.bin <- file.path(
+        basilisk.utils::getCondaDir(), 
+        basilisk.utils::getCondaBinary()
+    )
     conda_install(tmp2, packages=c("scipy==1.9.1"), conda=conda.bin)
 }
 
