### 2022-09-11 #4742 @LTLA
Thanks @jsun.

**(1)** Your comment prompted me to restructure the vignette to highlight the different pieces of information for different target audiences. Specifically, I made one section explicitly for package developers, and one section explicitly for end users. Admittedly this isn't close to what you suggested, but hear me out...

The (former) Section 3.4 was intended for package developers, and I didn't want to mention the environment variables at this point, because that would give the impression that package developers were expected to set environment variables... which they shouldn't, because that is the privilege/responsibility of the end user depending on the configuration of their R installation. Similarly, I didn't want to give end users the impression that they had to read through 3.4 and understand the intricacies of `basiliskRun()` stuff when all they want to do is to tweak some of the directory locations.

So, by marking things up more explicitly, it should hopefully be easier to understand who should do what. I suppose that, technically, I should distinguish between end users and R administrators, e.g., for installations on shared HPCs, though maybe the new vignette is clear enough as it is.

**(2)** Done. Added a truncated PCA from `sklearn.decomposition` and some stuff from `scipy.stats`.

Note that the example in the vignette is written for an end user and is a little different from the example in the manuscript (written for package developers). The set-up is currently a bit more involved; I will try to streamline it in a dedicated user-only function.

**(3)** Whoops. I think that part of the vignette must have been a legacy from the earliest versions of **basilisk**, when I allowed users to switch between `venv` and conda. IIRC it became increasingly difficult to reconcile as - I think - `venv` didn't allow use of different Python versions, and the Python package versions weren't exactly 1:1 between PyPI and conda. (Also some issues with Windows, but I forget the details.) In the end, I just gave up on the `venv` support and switched to conda for all systems.

In theory, it would be possible to modify a subset of the **basilisk** machinery to work for `venv`, though it would probably require a particularly skilled user to do it, as we'd be dependent on their Python and virtual environments being correctly configured. I don't think I could require client package developers to support this, as there's too many unknowns when the Python version is not pinned. If such users exist, I might consider generalizing `basiliskRun()` to work with `venv`.

---

These changes will land with **basilisk** 1.9.7, provided it passes through BioC's build system in a few days. Also added the explicit license text to accompany the mention in the `DESCRIPTION`.
