### 2022-09-12 #4742 @jsun
Thanks, @LTLA, sounds good! By the way, I checked the vignette hosted on the Bioconductor website last time; this time I downloaded the vignette source from the github main branch.

- For (1), I understood your opinion, as the main targets of basilisk are software developers;
- for (2), I can check the sample codes about `sklearn`, thanks (This is a great package, I'll use this in my research soon);
- and for (3), I see, I understood the situation.

By the way, it seems that I cannot build the vignette with my macOS, although I've already installed the latest basilisk and basilisk.utils. But this is not a critical issue, I'll finish my review soon (^^).
