### 2022-09-12 3a5d457
Create dedicated function for setting up environments for analysts.
