+++ vignettes/motivation.Rmd
@@ -234,18 +234,11 @@ This option only has an effect if `BASILISK_USE_SYSTEM_DIR` is set, as otherwise
 ## Using _basilisk_ directly for analyses
 
 While `r Biocpkg("basilisk")` is primarily intended for package developers, end users can also take advantage of its graceful handling of multiple Python environments in complex workflows.
-For example, we can easily instantiate a conda environment in our working directory with `setupBasiliskEnv()`:
+For example, we can easily instantiate a conda environment in our working directory with `createLocalBasiliskEnv()`:
 
 ```{r}
-# Environment location should be an absolute path.
-tmp <- file.path(getwd(), "basilisk-vignette-test") 
-
-# Avoid reinstantiating the environment if it already exists, for faster
-# re-execution of a workflow.
-if (!file.exists(tmp)) {
-    basilisk.utils::installConda() # just in case it hasn't been run before.
-    setupBasiliskEnv(tmp, packages=c("scikit-learn=1.1.1", "numpy=1.22.1"))
-}
+tmp <- createLocalBasiliskEnv("basilisk-vignette-test",
+    packages=c("scikit-learn=1.1.1", "numpy=1.22.1"))
 ```
 
 We can then supply this environment's path to `basiliskRun()` to execute Python-based calculations.
@@ -264,14 +257,14 @@ basiliskRun(env=tmp, fun=function(mat) {
 
 `basiliskRun()` can also be used with conda environments constructed outside of **basilisk**.
 Of course, in this case, it is the user's responsibility to ensure that the environment is correctly provisioned.
-Some caution is also required as `r Biocpkg("basilisk")` is not guaranteed to work with environments generated by a different version of conda, though problems seem to be rare in practice.
+Some caution is also required as `r Biocpkg("basilisk")` is not guaranteed to work with environments created by a different version of conda, though problems seem to be rare in practice.
 
 ```{r, error=FALSE, message=FALSE}
 library(reticulate)
 
-# Using reticulate directly to construct our conda environment; though we'll
-# cheat a little and use basilisk's conda installation, otherwise reticulate
-# will try to install its own miniconda.
+# In this case, we'll use reticulate directly to construct our conda
+# environment; though we'll cheat a little and use basilisk's conda
+# installation, otherwise reticulate will try to install its own miniconda.
 tmp2 <- file.path(getwd(), "basilisk-vignette-test2")
 if (!file.exists(tmp2)) {
     conda.bin <- file.path(
