+++ NAMESPACE
@@ -6,6 +6,7 @@ export(basiliskRun)
 export(basiliskStart)
 export(basiliskStop)
 export(configureBasiliskEnv)
+export(createLocalBasiliskEnv)
 export(findPersistentEnv)
 export(getBasiliskForceFallback)
 export(getBasiliskFork)
@@ -24,6 +25,7 @@ import(methods)
 importFrom(basilisk.utils,activateEnvironment)
 importFrom(basilisk.utils,getFallbackREnv)
 importFrom(basilisk.utils,getSystemDir)
+importFrom(basilisk.utils,installConda)
 importFrom(dir.expiry,lockDirectory)
 importFrom(dir.expiry,touchDirectory)
 importFrom(dir.expiry,unlockDirectory)
