### 2022-09-12 93466fa
Updated ignore files and README.
