### 2022-09-13 0439568
Made findPersistentEnv explicitly defunct.
