### 2022-09-14 33faab1
More general scipy trigger for fallback on BioC linux builders.
