### 2022-09-16 #4742 @gtonkinhill
Hi @LTLA,

I think this is a really useful package that significantly reduces the pain of combining R and python code. The paper is well written and I was able to successfully install and run the package. I have a few suggestions that are mostly things that I think would be nice to have rather than required changes.

1) It would be nice to include a function to automatically set up a new package to use `basilisk` similar to the `Rcpp.package.skeleton` function.
2) From the paper and the vignette, it was unclear to me how conda was installed and whether there was an easy method for updating the conda version. In my quick test run, it appeared to be using v4.8.3.
3) I was also wondering whether it would be possible to use the `mamba` solver instead of the default conda one. This can really speed things up for environments with a lot of dependencies.
4) I think it might be nice to include an example of selecting the python version for an environment in the example package and in the vignette. In fact, it might be preferable to suggest that users do this for every environment to ensure reproducibility. When testing the example package in `/inst/` my installation defaulted to using python 3.7 which led to version clashes between the python packages used in the example. Conda’s error messages are not particularly informative, so it took me a second to realise that it was the python version causing the issue.
5) You mention that the conda environments use a large amount of disk space. They can also generate a very large number of files which I have run into issues with when IT departments implement unnecessary restrictions. It would be great if there were a function to automatically clean up environments either by deleting them all or deleting those associated with a particular package.
