### 2022-09-18 6f11e33
Added utility to get the default Python version.
