+++ R/listCorePackages.R
@@ -11,10 +11,14 @@
 #'
 #' @author Aaron Lun
 #'
-#' @return A data.frame containing the \code{full}, a versioned package string, and \code{package}, the package name.
+#' @return 
+#' For \code{listPackages}, a data.frame containing the \code{full}, a versioned package string, and \code{package}, the package name.
+#'
+#' For \code{listPythonVersion}, a string containing the default version of Python.
 #' 
 #' @examples
 #' listPackages()
+#' listPythonVersion()
 #' 
 #' @export
 listPackages <- function(env=NULL) {
@@ -39,3 +43,18 @@ listCorePackages <- function() {
     .Deprecated(new="listPackages")
     listPackages()
 }
+
+#' @export
+#' @rdname listPackages
+listPythonVersion <- function(env=NULL) {
+    envpath <- obtainEnvironmentPath(env)
+    .python_version(envpath)
+}
+
+.python_version <- function(dir) {
+    py.cmd <- getPythonBinary(dir)
+    dump <- system2(py.cmd, "--version", stdout=TRUE, stderr=TRUE)
+    pattern <- "^Python "
+    sub(pattern, "", dump[grep(pattern, dump)[1]])
+}
+
