+++ NAMESPACE
@@ -13,6 +13,7 @@ export(getBasiliskFork)
 export(getBasiliskShared)
 export(listCorePackages)
 export(listPackages)
+export(listPythonVersion)
 export(obtainEnvironmentPath)
 export(setBasiliskForceFallback)
 export(setBasiliskFork)
