### 2022-09-18 6cd36db
Officially provide an option for devs to disable version checks.
