### 2022-09-18 7885fdb
Cleaned up setup docs.
