### 2022-09-18 2d59664
Updated vignette for Python version spec, new env vars.
