### 2022-09-18 da43437
Mention directory clearing functions in vignette.
