### 2022-09-18 #4742 @LTLA
Thanks @gtonkinhill.

**1.** An interesting idea. It would probably help anyone trying to build a **basilisk** client from scratch. I'd need to think about it though; at least a few of the existing clients come from the position of having an existing R package that interfaces with Python, and they only added **basilisk** integration as part of the BioC review process. In those cases, a more useful function would be to, e.g., `basiliskize()` an existing package by injecting the relevant bits and pieces into an existing R package. I'll have to see how much demand there is for that, maybe it would make the BioC review process easier.

**2.** The conda installation happens automatically the first time that `basiliskRun()` is called in any client package, so the user doesn't have to type anything special to get it - I added a note about this in the manuscript. There are a few exceptions, though:

- For a system installation, the conda installation is created during **basilisk**'s own installation
- If `BASILISK_EXTERNAL_CONDA` is set, that it just used directly without **basilisk** installing its own copy.

I also added a `BASILISK_MINICONDA_VERSION` environment variable that allows users to specify a different conda version, if they must. There are a few caveats though, which are mentioned in the updated vignette.

**3.** I suppose it is possible, though I think I would prefer to wait for mamba to become conda's default solver. Apparently this will happen at some point (see [here](https://www.anaconda.com/blog/a-faster-conda-for-a-growing-community)), so it should be fairly robust at that point, and I wouldn't have to take responsibility for any differences in behavior with the default.

**4.** Added some notes about this to the vignette and documentation.

It's worth noting that, actually, the Python version is explicitly pinned by `setupBasiliskEnv` to whatever was supplied with the conda installer, if no `python=` package string is detected in `packages=`. This is done as a fallback if client package developers forget to pin their Python versions. In your case, I'm guessing you used the BioC release version of **basilisk**, which pins Python to 3.7; but the `inst` example on GitHub uses the BioC devel version, which is pinned to 3.8, hence the issues.

**5.** Added some notes to the vignette about `basilisk.utils::clearExternalDir()`, which does exactly this.
