### 2022-10-10 #4742 @Nikoleta-v3
@LTLA thank you for your patience! I have no further comments. The software is very useful and well developed, and the manuscript is very well written.

The package is aimed at developers and thus it’s not very beginner friendly, but this is not an issue.

At this point could you please:

- [ ] (if necessary) Make a new tagged release of your software, and list the version tag of the archived version here. Or let me know if I should just use the release `1.9.2`
