### 2022-10-16 #4742 @LTLA
Thanks @Nikoleta-v3:

> * Make a new tagged release of your software, and list the version tag of the archived version here.

Done, 1.9.11 (https://github.com/LTLA/basilisk/releases/tag/1.9.11).

> * [ ]  Archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository)

Done, https://zenodo.org/record/7212841#.Y0uwA9LMLkw

DOI: 10.5281/zenodo.7212841
