### 2022-10-27 e185224
Don't let conda choose a different Python version in vignette.
