### 2022-10-31 c0b42ff
Test pandas load in the basiliskStart examples.
