+++ R/basiliskStart.R
@@ -159,7 +159,7 @@
 #' cl <- basiliskStart(tmploc)
 #' basiliskRun(proc=cl, function() { 
 #'     X <- reticulate::import("pandas"); X$`__version__` 
-#' })
+#' }, testload="pandas")
 #' basiliskStop(cl)
 #'
 #' # This happily co-exists with our other environment:
@@ -171,7 +171,7 @@
 #' cl2 <- basiliskStart(tmploc2)
 #' basiliskRun(proc=cl2, function() { 
 #'     X <- reticulate::import("pandas"); X$`__version__` 
-#' })
+#' }, testload="pandas")
 #' basiliskStop(cl2)
 #'
 #' # Persistence of variables is possible within a Start/Stop pair.
