### 2022-11-02 #4742 @Kevin-Mattheus-Moerman
@LTLA I have read your paper and studied the ZENODO archive. All seems in order as I have only a minor comment.

- [ ] Perhaps consider if a comma is needed after after `...This is burdensome, error-prone...`

Comment: This is fantastic :smile: `The name “basilisk” is based on the mythological snake monster (Rowling, 1998). The original purpose of the basilisk package was to freeze Python package versions, much like how the monster was able to Petrify its victims.`
