### 2024-02-10 #4742 @LTLA
@arfon As discussed in #5777, I've fixed my name on this manuscript. Note that the repo was recently transferred to the Bioconductor GitHub org; the `submission` branch is still present so all should be fine if GitHub's redirects are followed.
